# src/main/webapp/app/shared/api/generated/API-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "OncoKB™ is a precision oncology knowledge base developed at Memorial Sloan Kettering Cancer Center that contains biological and clinical information about genomic alterations in cancer.",
    "version": "v1.0.0",
    "title": "OncoKB™ APIs",
    "contact": {
      "name": "OncoKB™",
      "url": "https://www.oncokb.org",
      "email": "contact@oncokb.org"
    },
    "license": {
      "name": "Terms of Use",
      "url": "https://www.oncokb.org/terms"
    }
  },
...
  "paths": {
...
    "/api/companies/{id}": {
      "get": {
        "tags": [
          "company-resource"
        ],
        "summary": "getCompany",
        "operationId": "getCompanyUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CompanyDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "company-resource"
        ],
        "summary": "deleteCompany",
        "operationId": "deleteCompanyUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/usage/resources": {
      "get": {
        "tags": [
          "usage-analysis-controller"
        ],
        "summary": "resourceDetailGet",
        "operationId": "resourceDetailGetUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "description": "endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UsageSummary"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/usage/summary/resources": {
      "get": {
        "tags": [
          "usage-analysis-controller"
        ],
        "summary": "resourceUsageGet",
        "operationId": "resourceUsageGetUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UsageSummary"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/usage/summary/users": {
      "get": {
        "tags": [
          "usage-analysis-controller"
        ],
        "summary": "userOverviewUsageGet",
        "operationId": "userOverviewUsageGetUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "companyId",
            "in": "query",
            "description": "companyId",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserOverviewUsage"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/usage/users/{userId}": {
      "get": {
        "tags": [
          "usage-analysis-controller"
        ],
        "summary": "userUsageGet",
        "operationId": "userUsageGetUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserUsage"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "CompanyDTO": {
      "type": "object",
      "required": [
        "companyType",
        "licenseModel",
        "licenseStatus",
        "licenseType",
        "name"
      ],
      "properties": {
        "businessContact": {
          "type": "string"
        },
        "companyDomains": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "companyType": {
          "type": "string",
          "enum": [
            "PARENT",
            "BRANCH",
            "UNKNOWN"
          ]
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "legalContact": {
          "type": "string"
        },
        "licenseModel": {
          "type": "string",
          "enum": [
            "FULL",
            "LIMITED"
          ]
        },
        "licenseStatus": {
          "type": "string",
          "enum": [
            "TRIAL",
            "REGULAR",
            "TRIAL_EXPIRED",
            "EXPIRED",
            "UNKNOWN"
          ]
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "name": {
          "type": "string"
        },
        "numberOfUsers": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "CompanyDTO"
    },
...
    "UsageSummary": {
      "type": "object",
      "properties": {
        "day": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        },
        "month": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        },
        "year": {
          "type": "object",
          "additionalProperties": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "title": "UsageSummary"
    },
...
    "UserOverviewUsage": {
      "type": "object",
      "properties": {
        "endpoint": {
          "type": "string"
        },
        "maxUsage": {
          "type": "integer",
          "format": "int64"
        },
        "noPrivateEndpoint": {
          "type": "string"
        },
        "noPrivateMaxUsage": {
          "type": "integer",
          "format": "int64"
        },
        "totalUsage": {
          "type": "integer",
          "format": "int64"
        },
        "userEmail": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "UserOverviewUsage"
    },
    "UserUsage": {
      "type": "object",
      "properties": {
        "company": {
          "type": "string"
        },
        "jobTitle": {
          "type": "string"
        },
        "licenseType": {
          "type": "string"
        },
        "summary": {
          "$ref": "#/definitions/UsageSummary"
        },
        "userEmail": {
          "type": "string"
        },
        "userFirstName": {
          "type": "string"
        },
        "userLastName": {
          "type": "string"
        }
      },
      "title": "UserUsage"
    },
...
  }
}